Create mediainfo3.json
This is based on the existing mediainfo2.json example.  
Changes:
- move the example to a subdirectory based on the schema (preservation action)
- example shows how to run mediainfo directly on the command line, instead of using cwltool
- removed AnyOf input properties, limiting this example to a single input format
- remove output properties and replaced with raw output (ebucore xml)
- removed preservationActionName field (not a valid field i.e. not defined in preservation-action.json schema?)

Outstanding Issues:
This example still does not fully describe how to run this preservation action.  In order to do that, the par-parameter schema needs to be defined, so the command line parameters to mediainfo can be documented (these are described in the example field presently).

It is not clear what the 'value' field of rawOutputs should be.  In this example we assume that the value will eventually be an enumeration (xml, text, binary perhaps?).  We are assuming that 'value' means something like 'type' - describing what kind of raw output to expect.